Route post datasource in vault: merge 2 routes
